requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
email-validator
numpy
scipy
speechrecognition
pyaudio
